refactor: make `keys/mod.rs` compilable to wasm
