+++ parser/Cargo.toml
@@ -29,8 +29,3 @@ kanata-keyberon = { path = "../keyberon" }
 [features]
 cmd = []
 interception_driver = []
-
-[target.'cfg(target_os = "linux")'.dependencies]
-evdev = "=0.12.0"
-
-[target.'cfg(target_os = "windows")'.dependencies]
