+++ src/oskbd/mod.rs
@@ -9,3 +9,53 @@ pub use linux::*;
 mod windows;
 #[cfg(target_os = "windows")]
 pub use windows::*;
+
+// ------------------ KeyValue --------------------
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum KeyValue {
+    Release = 0,
+    Press = 1,
+    Repeat = 2,
+}
+
+impl From<i32> for KeyValue {
+    fn from(item: i32) -> Self {
+        match item {
+            0 => Self::Release,
+            1 => Self::Press,
+            2 => Self::Repeat,
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl From<bool> for KeyValue {
+    fn from(up: bool) -> Self {
+        match up {
+            true => Self::Release,
+            false => Self::Press,
+        }
+    }
+}
+
+impl From<KeyValue> for bool {
+    fn from(val: KeyValue) -> Self {
+        matches!(val, KeyValue::Release)
+    }
+}
+
+use kanata_parser::keys::OsCode;
+
+#[derive(Debug, Clone, Copy)]
+pub struct KeyEvent {
+    pub code: OsCode,
+    pub value: KeyValue,
+}
+
+#[cfg(not(all(feature = "interception_driver", target_os = "windows")))]
+impl KeyEvent {
+    pub fn new(code: OsCode, value: KeyValue) -> Self {
+        Self { code, value }
+    }
+}
